README.md
Hello 😃

Learning is very interesting. There are difficult tasks, but after you complete them, your mood rises.

Git was completely new to me.)

On the courses, I was surprised by the fact that it is possible to clone repositories. I was also surprised by how the Putty program works

So I will definitely use Linux in the future.




Thank you very much for the opportunity to study.
Good luck with your studies  to everyone.